Add validation for MangaChapter Controller

+++ Manga/MediatR/MangaChapter/Commands/Delete/DeleteChapterCommandHandler.cs
@@ -1,4 +1,5 @@
 ﻿using AutoMapper;
+using Manga.Exceptions;
 using Manga.Services.IServices;
 using MediatR;
 
@@ -15,6 +16,10 @@ namespace Manga.MediatR.MangaChapter.Commands.Delete
         public async Task Handle(DeleteChapterCommand request, CancellationToken cancellationToken)
         {
             var chapter = await _chapterService.GetByIdAsync(request.id);
+            if (chapter is null)
+            {
+                throw new NotFoundException($"There is no chapter with id {request.id}");
+            }
             await _chapterService.RemoveAsync(chapter);
         }
     }
